Split CPolyline and CPULineStyle to Renderer.h, make Hatch.cpp compile seperately
