+++ 62.CAD/CMakeLists.txt
@@ -5,7 +5,7 @@ if(NOT RES)
 endif()
 
 set(EXAMPLE_SOURCES
-  "${CMAKE_CURRENT_SOURCE_DIR}/curves.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/curves.cpp"
+  "${CMAKE_CURRENT_SOURCE_DIR}/Hatch.cpp"
 )
 nbl_create_executable_project("${EXAMPLE_SOURCES}" "" "" "" "${NBL_EXECUTABLE_PROJECT_CREATION_PCH_TARGET}")
